Updating the examples so that they run with numpy (backward compatibility). 1) Removing the solvers tests that rely on the proximal software, which is never properly tested in odl 2) Removing the simple_r and simple_rn examples, which don't make much sense 3) Removing the ray_trafo_parallel_3d_euler, which is never tested in ODL. 4) Removing the ufuncs examples 5) Modified a lot of examples to replace numpy calls on ODL objects by odl. calls.

+++ examples/tomo/backends/astra_performance_cuda_cone_3d_cg.py
@@ -24,7 +24,7 @@ det_size = 512
 niter = 10
 
 # Create reconstruction space
-reco_space = odl.uniform_discr(-domain_size / 2, domain_size / 2, domain_size)
+reco_space = odl.uniform_discr(-domain_size / 2, domain_size / 2, domain_size, dtype='float32')
 
 # Create geometry
 apart = odl.uniform_partition(0, 2 * np.pi, n_angles)
